LayoutTests/fast/dom/HTMLFormElement/htmlformelement-indexed-getter.html

<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>

<form id='f'>
  <input type="radio" name="a" value="a" id="i">HTMLInputElement</input>
</form>

<script>
description('This test verifies that elements can be fetched by index from HTMLFormElements.');

var form = document.getElementById('f');
var my_input = document.getElementById('i');
shouldBe("form[0]", "my_input");
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
